More fixes

+++ korge/src/commonTest/kotlin/TestExt.kt
@@ -1,6 +1,7 @@
+
+import korlibs.math.*
 import korlibs.math.geom.*
 import korlibs.math.geom.bezier.*
-import korlibs.math.isAlmostEquals
 import kotlin.test.*
 
 fun <
@@ -30,7 +31,7 @@ private fun <T : Any> T?.isAlmostEqualsGeneric(
     val e = this
     if (e == null || a == null) return (e == null) && (a == null)
     return when (e) {
-        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance.toFloat())
+        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance)
         is MPoint -> e.isAlmostEquals((a as? MPoint?) ?: return false, absoluteTolerance)
         is Vector3F -> e.isAlmostEquals((a as? Vector3F?) ?: return false, absoluteTolerance.toFloat())
         is Vector4F -> e.isAlmostEquals((a as? Vector4F?) ?: return false, absoluteTolerance.toFloat())
@@ -44,9 +45,9 @@ private fun <T : Any> T?.isAlmostEqualsGeneric(
             if (e.isNaN() && a.isNaN()) return true
             e.isAlmostEquals(a, absoluteTolerance)
         }
-        is Matrix -> e.isAlmostEquals((a as Matrix), absoluteTolerance.toFloat())
-        is MatrixTransform -> e.isAlmostEquals((a as MatrixTransform), absoluteTolerance.toFloat())
-        is Rectangle -> e.isAlmostEquals((a as Rectangle), absoluteTolerance.toFloat())
+        is Matrix -> e.isAlmostEquals((a as Matrix), absoluteTolerance)
+        is MatrixTransform -> e.isAlmostEquals((a as MatrixTransform), absoluteTolerance)
+        is Rectangle -> e.isAlmostEquals((a as Rectangle), absoluteTolerance)
         is MRectangle -> e.isAlmostEquals((a as MRectangle), absoluteTolerance)
         is Bezier -> e.points.isAlmostEqualsGeneric((a as? Bezier)?.points, absoluteTolerance)
         is PointList -> e.toList().isAlmostEqualsGeneric((a as? PointList)?.toList(), absoluteTolerance)
